+++ src/main/java/com/sandbox/domain/todos/dto/ErrorTodoResp.java
@@ -2,12 +2,8 @@ package com.sandbox.domain.todos.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
-@NoArgsConstructor
 @AllArgsConstructor
 public class ErrorTodoResp extends RuntimeException {
     private String message;
