+++ src/main/java/com/sandbox/domain/todos/entity/Todo.java
@@ -1,4 +1,4 @@
-package com.sandbox.domain.todos.dto;
+package com.sandbox.domain.todos.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
